Move <Layout> into its own directory and update affected files
